# .github/backend/backend/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "backend, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Reviews/submit": {
      "post": {
        "tags": [
          "Reviews"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ReviewDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Reviews/fetch-by-product": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "productId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reviews/remove/{reviewId}": {
      "delete": {
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "reviewId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reviews/search-reviews": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "username",
            "type": "string"
          },
          {
            "in": "query",
            "name": "orderId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ReviewDto": {
      "type": "object",
      "properties": {
        "reviewId": {
          "format": "int32",
          "type": "integer"
        },
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "reviewText": {
          "type": "string"
        },
        "productIds": {
          "type": "array",
          "items": {
            "format": "int32",
            "type": "integer"
          }
        }
      }
    }
  }
}